pyproject.toml
[project]
name = "avc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pdfplumber>=0.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "selenium>=4.35.0",
]

[tool.ruff]
include = ["*.py"]
src = ["avc"]
fix = true
indent-width = 4
line-length = 80

[tool.ruff.lint]
ignore = ["E402", "F401"]

[tool.basedpyright]
include = [".venv", "avc"]
verboseOutput = true
reportUnusedCallResult = false
reportMissingTypeStubs = false
reportAny = false
reportExplicitAny = false
reportUnknownParameterType = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportUnknownVariableType = false
reportUnknownMemberType = false
